pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Milyzway-Scanner"
version = "0.1.0"
description = "A modular and extensible vulnerability scanner."
readme = "README.md"
authors = [{name="critfinds", email="jsbtc1@proton.me"}]
requires-python = ">=3.6"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "aiohttp",
    "rich",
    "playwright",
    "PyYAML",
    "beautifulsoup4",
    "python-dotenv",
    "pydantic>=2.0.0",
]
urls = {Homepage = "https://github.com/critfinds/Milyzway-Scanner"}

[project.scripts]
milyzway-scanner = "scanner.app:main"

[tool.setuptools.packages.find]
include = ["scanner*"]